+++ lib/PingService/AbstractPingUtils.php
@@ -8,22 +8,41 @@ use Symfony\Component\HttpFoundation\Response;
 
 abstract class AbstractPingUtils
 {
-    public static function ping(string $path): array
+    /** @var string */
+    private $path;
+    /** @var bool */
+    private $is_dev;
+    /** @var array */
+    private $options;
+
+    public static function ping(string $path, array $options = [], bool $is_dev = false): array
     {
-        if (\Config::$UNDER_DEV) {
+        return (new static($path, $options, $is_dev))->pingWithRetry();
+    }
+
+    private function __construct(string $path, array $options, bool $is_dev)
+    {
+        $this->path = $path;
+        $this->options = $options;
+        $this->is_dev = $is_dev;
+    }
+
+    private function pingWithRetry(): array
+    {
+        if ($this->is_dev) {
             return ['http_code' => 0, 'error' => 'under_dev', 'body' => false];
         }
 
-        $url = static::getBaseUrl() . $path;
+        $url = $this->getBaseUrl() . $this->path;
 
         $retry_count = 0;
         $result_dict = [];
-        while ($retry_count <= static::getMaxRetryCount()) {
-            $result_dict = self::getContents($url);
+        while ($retry_count <= $this->getMaxRetryCount()) {
+            $result_dict = $this->getContents($url);
             if ($result_dict['http_code'] === Response::HTTP_OK) {
                 return $result_dict;
             }
-            sleep(static::getRetryWaitSeconds());
+            sleep($this->getRetryWaitSeconds());
             $retry_count++;
         }
 
@@ -33,16 +52,16 @@ abstract class AbstractPingUtils
     /**
      * @param string $url
      *
-     * @return array 호출 결과 [http_code, error, body]
+     * @return array 호출 결과 [http_code, body]
      */
-    private static function getContents(string $url): array
+    private function getContents(string $url): array
     {
         $option = [
-            'timeout' => static::getTimeoutSeconds(),
-            'connect_timeout' => static::getConnectionTimeoutSeconds(),
+            'timeout' => $this->getTimeoutSeconds(),
+            'connect_timeout' => $this->getConnectionTimeoutSeconds(),
         ];
 
-        $client = new Client($option);
+        $client = new Client(array_merge($this->options, $option));
 
         try {
             $response = $client->get($url);
@@ -56,24 +75,24 @@ abstract class AbstractPingUtils
         return ['http_code' => $status_code, 'body' => $body];
     }
 
-    abstract protected static function getBaseUrl(): string;
+    abstract protected function getBaseUrl(): string;
 
-    protected static function getMaxRetryCount(): int
+    protected function getMaxRetryCount(): int
     {
         return 10;
     }
 
-    protected static function getRetryWaitSeconds(): int
+    protected function getRetryWaitSeconds(): int
     {
         return 6;
     }
 
-    protected static function getTimeoutSeconds(): int
+    protected function getTimeoutSeconds(): int
     {
         return 10;
     }
 
-    protected static function getConnectionTimeoutSeconds(): int
+    protected function getConnectionTimeoutSeconds(): int
     {
         return 10;
     }