+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- Remove `\Config` reference in `PingService`
+
 ## [1.0.8] - 2020-04-28
 ### Fixed
 - Fixed null|string type error when variable $values is empty string or null